moved version stuff to own module
