+++ pages/electron/index.vue
@@ -209,11 +209,13 @@
       ]),
       ...mapGetters('electron', [
         'updateAvailable',
-        'visibleReleases',
-        'latestVisibleVersion',
         'currentSeedPathBasename',
         'isNewVersion',
       ]),
+      ...mapGetters('version', [
+        'latestVisibleVersion',
+        'visibleReleases',
+      ])
     },
     watch: {
       currentVersion: {
